pyproject.toml
[project]
name = "active-reading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "datasets>=4.4.1",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "wikipedia>=1.4.0",
]
